% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusionStats.R
\name{analisis_sobrevida}
\alias{analisis_sobrevida}
\title{analisis_sobrevida}
\usage{
analisis_sobrevida(stats, metadata, thre = "median")
}
\arguments{
\item{stats}{dataframe that contains}

\item{metadata}{must be a name of one of the columns of the metadata}

\item{thre}{string}
}
\description{
Generates
}
