% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FusionAnnotation.R
\name{AnotateFusions}
\alias{AnotateFusions}
\title{AnotateFusions}
\usage{
AnotateFusions(protein_df, fusions_report)
}
\arguments{
\item{protein_df}{df}

\item{fusions_report}{df}
}
\description{
Generates
}
