% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusionStats.R
\name{boxplots_TFB_MTT}
\alias{boxplots_TFB_MTT}
\title{boxplots_TFB_MTT.}
\usage{
boxplots_TFB_MTT(
  stats,
  group,
  cohorte,
  Apareados = FALSE,
  grupo_ids_apareados = NA
)
}
\arguments{
\item{stats}{dataframe that contains}

\item{group}{must be a name of one of the columns of the metadata}

\item{cohort}{string}
}
\description{
Generates
}
