% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{downloadArriba}
\alias{downloadArriba}
\title{downloadArriba}
\usage{
downloadArriba(soft_directory)
}
\value{
The path where the .exe file is located
}
\description{
Downloads and decompresses the GATK software
}
