% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadHG38.R
\name{downloadHG38}
\alias{downloadHG38}
\title{downloadHG38}
\usage{
downloadHG38(soft_directory)
}
\value{
Paths of FASTA and GTF(annotation) from the genome reference.
}
\description{
Downloads the FASTA and the annotation of the genome reference version HG38.
}
