% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCReport-Grupal.R
\name{generateCohortQCreport}
\alias{generateCohortQCreport}
\title{generateCohortQCreport}
\usage{
generateCohortQCreport(
  patients_dir,
  bbmap,
  path_bbmap,
  index_rsubread,
  rate,
  rmd_content
)
}
\arguments{
\item{patients_dir}{path were the folders of the samplesare stored.}

\item{bbmap}{TRUE/FALSE}

\item{path_bbmap}{path to bbmap}

\item{index_rsubread}{path to the index of the mycoplasma}

\item{rate}{"0.6", from 0 to 1.}

\item{rmd_content}{generated before}
}
\description{
kvjdfnkvjdf
}
