% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runARRIBA.R
\name{generateFusionsReport}
\alias{generateFusionsReport}
\title{generate Fusions Report}
\usage{
generateFusionsReport(fusion.file, assemblyVersion, patient_id, patient_dir)
}
\arguments{
\item{fusion.file}{Path of the fusions.tsv file generated by ARRIBA.}

\item{assemblyVersion}{version of the genome reference}

\item{patient_id}{string with patient identification}

\item{patient_dir}{path of the patient's folder.}
}
\description{
Generates an xlsx file with the information of the fusions detected by ARRIBA.
}
