% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCReport-Individual.R
\name{generateQCreport_Individual}
\alias{generateQCreport_Individual}
\title{Generate QC report}
\usage{
generateQCreport_Individual(patient_dir)
}
\arguments{
\item{patients_dir}{fskbfjsf}
}
\description{
kvjdfnkvjdf
}
