% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadHG38.R
\name{indexRefSTAR}
\alias{indexRefSTAR}
\title{index Reference genome with STAR}
\usage{
indexRefSTAR(AnnotationHG38, FastaHG38)
}
\value{
Paths of FASTA and GTF(annotation) from the genome reference.
}
\description{
Index genome reference version HG38 with STAR.
}
