% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runARRIBA.R
\name{runARRIBA}
\alias{runARRIBA}
\title{RunARRIBA}
\usage{
runARRIBA(patient_dir, genomeversion = "hg38", assemblyVersion = "GRCh38")
}
\arguments{
\item{patient_dir}{Path of the directory that contains bam file generated by STAR}
}
\description{
Executes the Arriba software from the bam file generated by STAR.
}
