% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFastQC.R
\name{runFastQC}
\alias{runFastQC}
\title{Run FastQC}
\usage{
runFastQC(patient_dir)
}
\arguments{
\item{patent_dir}{Path of the directory that contains R1 and R2 fastq files. It can be either the "trimmed" folder or the original folder.}
}
\description{
Executes the FastQC for R1 and R2.
}
