% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFeatureCounts.R
\name{runFeatureCounts}
\alias{runFeatureCounts}
\title{Run Feature Counts}
\usage{
runFeatureCounts(patients_dir, genomeRef)
}
\arguments{
\item{patients_dir}{directory that contains}
}
\value{
FC.object featureCounts object.
}
\description{
Executes the feature count function from Rsubread library.
}
