% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSTAR.R
\name{runSTAR}
\alias{runSTAR}
\title{RunSTAR}
\usage{
runSTAR(patient_dir, twoPass = c("None", "Basic"), soft_directory)
}
\arguments{
\item{patient_dir}{Path of the directory that contains R1 and R2 fastq files. It can be either the "trimmed" folder or the original folder.}

\item{twoPass}{kjvdfk}

\item{nThreads}{vlkdnf}
}
\value{
path of the BAM generated by STAR
}
\description{
Executes the alignment of the sample with STAR software.
}
