% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTrimgalore.R
\name{runTrimgalore}
\alias{runTrimgalore}
\title{RunTrimgalore}
\usage{
runTrimgalore(patient_dir, trim_quality = 30, soft_directory)
}
\arguments{
\item{patient_dir}{Path donde se encuentra el archivo R1 de formato fasta o fastq.}

\item{trim_quality}{is the minimum value of the quality of each base within the sequence that will pass the filter.}
}
\value{
path of the trimmed folder which contained the trimmed files and was created inside the patient folder.
}
\description{
Corre la funcion TrimGalore para eliminar los adapters y las lecturas (y bases) de baja calidad.
}
